Subida entidades cambio llaves integer DANIEL

+++ ppp/src/main/java/com/api/ppp/back/models/Actividad.java
@@ -1,10 +1,12 @@
 package com.api.ppp.back.models;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
 import java.io.Serializable;
+import java.util.List;
 
 @Getter
 @Setter
@@ -15,7 +17,7 @@ public class Actividad implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "act_id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "act_descripcion")
     private String descripcion;
@@ -23,8 +25,6 @@ public class Actividad implements Serializable {
     @OneToOne
     @JoinColumn(name = "sol_id")
     private SolicitudEmpresa solicitudEmpresa;
-
-
 
 
 }
